se cambio el MaestroDTO
cambio el Long por int

+++ DTOs/src/main/java/dtos/MaestroDTO.java
@@ -10,40 +10,46 @@ import java.util.List;
  *
  * @author joel_
  */
-public class MaestroDTO extends UsuarioDTO{
+public class MaestroDTO extends UsuarioDTO {
+
     protected List<String> materias;
-    
+
     /**
      * Constructor por omision
      */
-    public MaestroDTO(){}
+    public MaestroDTO() {
+    }
+
     /**
-     * Constructor que establece la lista de materias y tambien establece 
+     * Constructor que establece la lista de materias y tambien establece
      * atributos de la clase padre (UsuarioDTO)
+     *
      * @param materias
      * @param nombre
      * @param correo
      * @param id
-     * @param contrasenia 
+     * @param contrasenia
      */
-    public MaestroDTO(List<String> materias, String nombre, String correo, Long id, String contrasenia) {
+    public MaestroDTO(List<String> materias, String nombre, String correo, int id, String contrasenia) {
         super(nombre, correo, id, contrasenia);
         this.materias = materias;
     }
-    
+
     /**
      * Metodo para obtener la lista de materias
+     *
      * @return materias
      */
-    public List<String> getMateria(){
+    public List<String> getMateria() {
         return materias;
     }
-    
+
     /**
      * Metodo para establecer materias nuevas apartir de una lista nueva
-     * @param materias 
+     *
+     * @param materias
      */
-    public void setMateria(List<String> materias){
-        this.materias=materias;
+    public void setMateria(List<String> materias) {
+        this.materias = materias;
     }
 }
